Update mpi utility to allow for custom function which defines what task data is passed to workers
Previously workers were always just passed an index which was used to get task data from a
set list which was present on all workers. This can still be accomplished but now only
requires the list to be present on master, while also accommodating much more complex
processes. Most importantly, this now happens on demand when a worker requests a task,
so if the old task list required pulling multiple items from a database in bulk, they
can now be pulled one at a time as needed.